requirements.txt
cachetools==5.0.0
click==8.0.3
Flask==2.0.2
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
gspread==5.1.1
gspread-dataframe==3.2.2
httplib2==0.20.4
pandas==1.3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.8
uritemplate==4.1.1
Werkzeug==2.0.2
